Added end_at column to events table

+++ database/migrations/2020_06_20_140932_create_event_schedule_event_speaker_table.php
@@ -15,7 +15,7 @@ class CreateEventScheduleEventSpeakerTable extends Migration
     {
         Schema::create('event_schedule_event_speaker', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('event_speaker_id')->unsigned()->nullable();
+            $table->bigInteger('event_speaker_id')->unsigned()->nullable();
             $table->foreign('event_speaker_id')->references('id')->on('event_speakers');
             $table->integer('event_schedule_id')->unsigned()->nullable();
             $table->foreign('event_schedule_id')->references('id')->on('event_schedules');
